refactor: 임베디드 타입 명시
